requirements.txt
numpy==1.24.3
Pillow==9.5.0
scipy==1.10.1
